refactor: move root TS files to bin/, server/, and src/tui/
