+++ package.json
@@ -11,15 +11,15 @@
   "license": "MIT",
   "type": "module",
   "scripts": {
-    "dev:api": "bun run server.ts",
+    "dev:api": "bun run server/index.ts",
     "dev:ui": "vite --logLevel warn",
     "dev": "bun run --bun concurrently \"bun run dev:api\" \"bun run dev:ui\"",
     "build": "vite build",
     "build:assets": "bun run scripts/embed-dist.ts",
-    "build:binary": "vite build && bun run scripts/embed-dist.ts && bun build --compile cli.ts --outfile ./release/agentop",
+    "build:binary": "vite build && bun run scripts/embed-dist.ts && bun build --compile bin/cli.ts --outfile ./release/agentop",
     "preview": "vite preview",
-    "watch": "bun run watcher.ts",
-    "watch:cli": "bun run watch-cli.ts",
+    "watch": "bun run server/otel-watcher.ts",
+    "watch:cli": "bun run src/tui/index.ts",
     "test": "bun test",
     "prepare": "husky"
   },
